CI: move fetch-tor-debs job to weekly
deb.torproject.org no longer mirrors 3.5.x xenial packages, causing the fetch-tor-debs target to fail. Since we will be remaining on the 3.5.x series since it will outlive 4.0.x due to its LTS status, we should move this check to weekly as it is expected to fail until a new 3.5.x package is released.